Fix Ubah Dokumen SPP

+++ app/Http/Controllers/dashboard/spp/SppGuController.php
@@ -32,7 +32,7 @@ class SppGuController extends Controller
         $biroOrganisasi = in_array($role, ['Admin', 'PPK', 'ASN Sub Bagian Keuangan', 'Kuasa Pengguna Anggaran']) ? $request->biro_organisasi_id : Auth::user()->profil->biro_organisasi_id;
         if ($request->ajax()) {
             $data = SppGu::where(function ($query) use ($request, $biroOrganisasi, $role) {
-                if ($request->biro_organisasi_id && $request->biro_organisasi_id != 'Semua') {
+                if ($biroOrganisasi && $biroOrganisasi != 'Semua') {
                     $query->where('biro_organisasi_id', $biroOrganisasi);
                 }
 
@@ -140,9 +140,9 @@ class SppGuController extends Controller
                         $actionBtn .= '<a href="' . url('/surat-pernyataan/spp-gu/' . $row->id) . '" class="btn btn-success btn-sm mr-1"><i class="fas fa-envelope"></i> Surat Pernyataan</a>';
                     }
 
-                    if (in_array(Auth::user()->role, ["Admin", "Bendahara Pengeluaran"])) {
+                    if (in_array(Auth::user()->role, ['Admin', 'Bendahara Pengeluaran', 'Bendahara Pengeluaran Pembantu', 'Bendahara Pengeluaran Pembantu Belanja Hibah'])) {
                         $actionBtn .= '<a class="btn btn-primary text-light btn-sm mr-1" href="' . url('spp-gu/' . $row->id) . '"><i class="far fa-check-circle"></i> Lihat</a>';
-                        if (($row->status_validasi_akhir == 0 && Auth::user()->role == "Bendahara Pengeluaran") || Auth::user()->role == "Admin") {
+                        if (($row->status_validasi_akhir == 0 && in_array(Auth::user()->role, ['Admin', 'Bendahara Pengeluaran', 'Bendahara Pengeluaran Pembantu', 'Bendahara Pengeluaran Pembantu Belanja Hibah']) && $row->biro_organisasi_id == Auth::user()->profil->biro_organisasi_id) || Auth::user()->role == "Admin") {
                             $actionBtn .= '<button id="btn-delete" class="btn btn-danger btn-sm mr-1" value="' . $row->id . '" > <i class="fas fa-trash-alt"></i> Hapus</button>';
                         }
                     }
@@ -224,7 +224,13 @@ class SppGuController extends Controller
         $jumlahAnggaran = 'Rp. ' . number_format($jumlahAnggaran->jumlah_anggaran, 0, ',', '.');
         $anggaranDigunakan = 'Rp. ' . number_format($sppGu->anggaran_digunakan, 0, ',', '.');
 
-        return view('dashboard.pages.spp.sppGu.createTahapAkhir', compact(['daftarTahun', 'daftarDokumenSppGu', 'sppGu', 'jumlahAnggaran', 'jumlahAnggaranHitung', 'anggaranDigunakan']));
+        $role = Auth::user()->role;
+
+        if (((in_array($role, ['Admin', 'PPK', 'ASN Sub Bagian Keuangan', 'Kuasa Pengguna Anggaran'])) || Auth::user()->profil->biro_organisasi_id == $sppGu->biro_organisasi_id) && ($sppGu->status_validasi_asn == 1 && $sppGu->status_validasi_ppk == 1 && $sppGu->tahap == "Awal")) {
+            return view('dashboard.pages.spp.sppGu.createTahapAkhir', compact(['daftarTahun', 'daftarDokumenSppGu', 'sppGu', 'jumlahAnggaran', 'jumlahAnggaranHitung', 'anggaranDigunakan']));
+        } else {
+            abort(403, 'Anda tidak memiliki akses halaman tersebut!');
+        }
     }
 
     /**
@@ -387,7 +393,12 @@ class SppGuController extends Controller
             $daftarDokumenSppGu = DokumenSppGu::where('spp_gu_id', $sppGu->id)->get();
         }
 
-        return view('dashboard.pages.spp.sppGu.show', compact(['sppGu', 'tipe', 'jumlahAnggaran', 'anggaranDigunakan', 'tahap', 'daftarDokumenSppGu']));
+        $role = Auth::user()->role;
+        if ((in_array($role, ['Admin', 'PPK', 'ASN Sub Bagian Keuangan', 'Kuasa Pengguna Anggaran'])) || Auth::user()->profil->biro_organisasi_id == $sppGu->biro_organisasi_id) {
+            return view('dashboard.pages.spp.sppGu.show', compact(['sppGu', 'tipe', 'jumlahAnggaran', 'anggaranDigunakan', 'tahap', 'daftarDokumenSppGu']));
+        } else {
+            abort(403, 'Anda tidak memiliki akses halaman tersebut!');
+        }
     }
 
     /**
@@ -414,7 +425,12 @@ class SppGuController extends Controller
             $daftarDokumenSppGu = DokumenSppGu::where('spp_gu_id', $sppGu->id)->get();
         }
 
-        return view('dashboard.pages.spp.sppGu.edit', compact(['sppGu', 'request', 'jumlahAnggaran', 'anggaranDigunakan', 'jumlahAnggaranHitung', 'daftarDokumenSppGu', 'tahap']));
+        $role = Auth::user()->role;
+        if (($role == "Admin" || Auth::user()->profil->biro_organisasi_id == $sppGu->biro_organisasi_id) && ($sppGu->status_validasi_asn == 2 || $sppGu->status_validasi_ppk == 2)) {
+            return view('dashboard.pages.spp.sppGu.edit', compact(['sppGu', 'request', 'jumlahAnggaran', 'anggaranDigunakan', 'jumlahAnggaranHitung', 'daftarDokumenSppGu', 'tahap']));
+        } else {
+            abort(403, 'Anda tidak memiliki akses halaman tersebut!');
+        }
     }
 
     /**
@@ -566,12 +582,21 @@ class SppGuController extends Controller
     {
         $sppGu->delete();
 
+        $riwayatSppGu = RiwayatSppGu::where('spp_gu_id', $sppGu->id)->whereNotNull('surat_penolakan')->get();
+        if (count($riwayatSppGu) > 0) {
+            foreach ($riwayatSppGu as $riwayat) {
+                Storage::delete('surat_penolakan_spp_gu/' . $riwayat->surat_penolakan);
+            }
+        }
+
         RiwayatSppGu::where('spp_gu_id', $sppGu->id)->delete();
 
         $dokumenSppGu = DokumenSppGu::where('spp_gu_id', $sppGu->id)->get();
-        foreach ($dokumenSppGu as $dokumen) {
-            Storage::delete('dokumen_spp_gu/' . $dokumen->dokumen);
-            $dokumen->delete();
+        if (count($dokumenSppGu) > 0) {
+            foreach ($dokumenSppGu as $dokumen) {
+                Storage::delete('dokumen_spp_gu/' . $dokumen->dokumen);
+                $dokumen->delete();
+            }
         }
 
         return response()->json(['status' => 'success']);
@@ -650,6 +675,12 @@ class SppGuController extends Controller
     {
         $tipeSuratPenolakan = 'spp-gu';
         $tipeSuratPengembalian = 'spp_gu';
-        return view('dashboard.pages.spp.sppGu.riwayat', compact(['sppGu', 'tipeSuratPenolakan', 'tipeSuratPengembalian']));
+
+        $role = Auth::user()->role;
+        if ((in_array($role, ['Admin', 'PPK', 'ASN Sub Bagian Keuangan', 'Kuasa Pengguna Anggaran'])) || Auth::user()->profil->biro_organisasi_id == $sppGu->biro_organisasi_id) {
+            return view('dashboard.pages.spp.sppGu.riwayat', compact(['sppGu', 'tipeSuratPenolakan', 'tipeSuratPengembalian']));
+        } else {
+            abort(403, 'Anda tidak memiliki akses halaman tersebut!');
+        }
     }
 }
